+++ README.md
@@ -1,4 +1,2 @@
 # flux-k8s
 Project to manage Flux tasks needed to standardize kubernetes HPC scheduling interfaces
-
-Code under `golang` directory is under Apache v2.0 license.
