--- examples/grpc-client/build.sh
@@ -1,1 +1,0 @@
-go build -o client main.go
