--- examples/grpc-client/run.sh
@@ -1,1 +1,0 @@
-./client
