--- examples/pi/clean_pods.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-kubectl delete -f ./pi-job-kubeflux-segfault.yaml
-kubectl delete -f ./pi-job-kubeflux.yaml
