--- examples/pi/demo_failed_pod_cancellation.sh
@@ -1,6 +1,0 @@
-#!/bin/sh
-
-kubectl create -f ./pi-job-kubeflux-segfault.yaml
-echo "wait 5 seconds"
-sleep 5
-kubectl create -f ./pi-job-kubeflux.yaml
