--- examples/pi/pi-job-default.yaml
@@ -1,20 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: pi-job-default-sched
-spec:
-  completions: 4
-  template:
-    metadata:
-      labels:
-        app: pi-test-default
-    spec:
-      containers:
-      - name: pi-test
-        image: quay.io/eduardoarango/pi:ubi8
-        resources:
-          limits:
-            cpu: "12"
-          requests:
-            cpu: "12"
-      restartPolicy: Never
