--- examples/pi/pi-job-kubeflux.yaml
@@ -1,21 +1,0 @@
-apiVersion: batch/v1
-kind: Job
-metadata:
-  name: pi-job-kubeflux-sched
-spec:
-  completions: 4
-  template:
-    metadata:
-      labels:
-        app: pi-test-kubeflux
-    spec:
-      schedulerName: scheduling-plugin # comment this out to get it scheduled by the default scheduler
-      containers:
-      - name: pi-test
-        image: quay.io/eduardoarango/pi:ubi8
-        resources:
-          limits:
-            cpu: "8"
-          requests:
-            cpu: "8"
-      restartPolicy: Never
