--- flux-plugin/kubeflux/go.mod
@@ -1,44 +1,0 @@
-module kubeflux
-
-go 1.16
-
-require (
-	google.golang.org/genproto v0.0.0-20210602131652-f16073e35f0c // indirect
-	google.golang.org/grpc v1.38.0
-	gopkg.in/yaml.v2 v2.4.0
-	k8s.io/api v0.22.3
-	k8s.io/apimachinery v0.22.3
-	k8s.io/client-go v0.22.3
-	k8s.io/kubernetes v1.22.3
-	fluxcli v0.0.0
-)
-
-replace (
-	fluxcli => /go/src/fluxcli
-	k8s.io/api => k8s.io/api v0.22.3
-	k8s.io/apiextensions-apiserver => k8s.io/apiextensions-apiserver v0.22.3
-	k8s.io/apimachinery => k8s.io/apimachinery v0.22.3
-	k8s.io/apiserver => k8s.io/apiserver v0.22.3
-	k8s.io/cli-runtime => k8s.io/cli-runtime v0.22.3
-	k8s.io/client-go => k8s.io/client-go v0.22.3
-	k8s.io/cloud-provider => k8s.io/cloud-provider v0.22.3
-	k8s.io/cluster-bootstrap => k8s.io/cluster-bootstrap v0.22.3
-	k8s.io/code-generator => k8s.io/code-generator v0.22.3
-	k8s.io/component-base => k8s.io/component-base v0.22.3
-	k8s.io/component-helpers => k8s.io/component-helpers v0.22.3
-	k8s.io/controller-manager => k8s.io/controller-manager v0.22.3
-	k8s.io/cri-api => k8s.io/cri-api v0.22.3
-	k8s.io/csi-translation-lib => k8s.io/csi-translation-lib v0.22.3
-	k8s.io/kube-aggregator => k8s.io/kube-aggregator v0.22.3
-	k8s.io/kube-controller-manager => k8s.io/kube-controller-manager v0.22.3
-	k8s.io/kube-proxy => k8s.io/kube-proxy v0.22.3
-	k8s.io/kube-scheduler => k8s.io/kube-scheduler v0.22.3
-	k8s.io/kubectl => k8s.io/kubectl v0.22.3
-	k8s.io/kubelet => k8s.io/kubelet v0.22.3
-	k8s.io/kubernetes => k8s.io/kubernetes v1.22.3
-	k8s.io/legacy-cloud-providers => k8s.io/legacy-cloud-providers v0.22.3
-	k8s.io/metrics => k8s.io/metrics v0.22.3
-	k8s.io/mount-utils => k8s.io/mount-utils v0.22.3
-	k8s.io/pod-security-admission => k8s.io/pod-security-admission v0.22.3
-	k8s.io/sample-apiserver => k8s.io/sample-apiserver v0.22.3
-)
